requirements.txt
fastapi[standard]
uvicorn
python-dotenv
pydantic
pydantic-settings
alembic
aiofiles
python-multipart
sqlmodel
ultralytics
pillow